+++ crates/approx-chol/src/approx_chol/factorization/approximate.rs
@@ -5,7 +5,7 @@ mod clique_tree;
 mod ordering;
 mod star;
 
-pub use clique_tree::clique_tree_sample;
+pub use clique_tree::CliqueTreeSampler;
 
 #[cfg(any(feature = "serde", test))]
 use super::FactorError;